# Optimize `TypeMatcher` data structure.